requirements.txt
streamlit==1.31.0
paddlepaddle==2.6.0
paddleocr==2.7.3
Pillow==10.2.0
openai==1.12.0
requests==2.31.0
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
